Rename constraint folder
