+++ .gitignore
@@ -16,8 +16,7 @@ zen_garden_env/
 .coverage
 *runscript.py
 *test.py
-/tests/testcases/test_1k__operation/*
-/tests/testcases/test_7b__operation/*
+*__operation*
 
 # pytest monitor
 .pymon
