+++ zen_garden/elements/carrier/constraints/__init__.py
@@ -1,6 +1,6 @@
 """Carrier constraints."""
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 from .availability_import_export_constraint import (
     AvailabilityImportExportConstraint,
