+++ zen_garden/elements/carrier/constraints/carbon_emission_carrier_constraint.py
@@ -1,4 +1,4 @@
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class CarbonEmissionsCarrierConstraint(GenericConstraint):
