+++ zen_garden/elements/carrier/constraints/carbon_emissions_carrier_total_constraint.py
@@ -1,4 +1,4 @@
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class CarbonEmissionsCarrierTotalConstraint(GenericConstraint):
