+++ zen_garden/elements/carrier/constraints/cost_limit_shed_demand_constraint.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class CostLimitShedDemandConstraint(GenericConstraint):
