+++ zen_garden/elements/carrier/constraints/nodal_energy_balance_constraint.py
@@ -3,9 +3,9 @@ import numpy as np
 import xarray as xr
 from linopy.expressions import LinearExpression
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
 from zen_garden.elements.carrier import Carrier
 from zen_garden.model.components.multi_index_helper import MultiIndexHelper
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class NodalEnergyBalanceConstraint(GenericConstraint):
