+++ zen_garden/elements/conversion_technology/constraints/capacity_factor_conversion_constraint.py
@@ -1,6 +1,6 @@
 import xarray as xr
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class CapacityFactorConversionConstraint(GenericConstraint):
