+++ zen_garden/elements/conversion_technology/constraints/carrier_conversion_constraint.py
@@ -5,7 +5,7 @@ import pandas as pd
 import xarray as xr
 from linopy.expressions import LinearExpression
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 from zen_garden.utils import align_like
