+++ zen_garden/elements/conversion_technology/constraints/linear_capex_constraint.py
@@ -2,7 +2,7 @@ import linopy as lp
 import numpy as np
 from linopy.expressions import LinearExpression
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class LinearCapexConstraint(GenericConstraint):
