+++ zen_garden/elements/conversion_technology/constraints/minimum_full_load_hours_constraint.py
@@ -1,7 +1,7 @@
 import numpy as np
 import xarray as xr
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class MinimumFullLoadHoursConstraint(GenericConstraint):
