+++ zen_garden/elements/conversion_technology/constraints/opex_emissions_technology_conversion_constraint.py
@@ -1,4 +1,4 @@
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class OpexEmissionsTechnologyConversionConstraint(GenericConstraint):
