+++ zen_garden/elements/energy_system/constraints/carbon_emissions_annual_limit_constraint.py
@@ -1,4 +1,4 @@
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class CarbonEmissionsAnnualLimitConstraint(GenericConstraint):
