+++ zen_garden/elements/energy_system/constraints/carbon_emissions_annual_overshoot_constraint.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class CarbonEmissionsAnnualOvershootConstraint(GenericConstraint):
