+++ zen_garden/elements/energy_system/constraints/carbon_emissions_cumulative_constraint.py
@@ -1,6 +1,6 @@
 import xarray as xr
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class CarbonEmissionsCumulativeConstraint(GenericConstraint):
