+++ zen_garden/elements/energy_system/constraints/net_present_cost_constraint.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class NetPresentCostConstraint(GenericConstraint):
