+++ zen_garden/elements/retrofitting_technology/constraints/__init__.py
@@ -1,6 +1,6 @@
 """Retrofitting technology constraints."""
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 from .retrofit_flow_coupling_constraint import (
     RetrofitFlowCouplingConstraint,
