+++ zen_garden/elements/retrofitting_technology/constraints/retrofit_flow_coupling_constraint.py
@@ -5,7 +5,7 @@ import logging
 
 import pandas as pd
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 from zen_garden.utils import align_like
 
 logger = logging.getLogger(__name__)
