+++ zen_garden/elements/storage_technology/constraints/__init__.py
@@ -1,6 +1,6 @@
 """Storage technology constraints."""
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 from .capacity_energy_to_power_ratio_constraint import (
     CapacityEnergyToPowerRatioConstraint,
