+++ zen_garden/elements/storage_technology/constraints/capacity_energy_to_power_ratio_constraint.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class CapacityEnergyToPowerRatioConstraint(GenericConstraint):
