+++ zen_garden/elements/storage_technology/constraints/capacity_factor_storage_constraint.py
@@ -1,6 +1,6 @@
 import xarray as xr
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class CapacityFactorStorageConstraint(GenericConstraint):
