+++ zen_garden/elements/storage_technology/constraints/couple_storage_level_constraint.py
@@ -1,4 +1,4 @@
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class CoupleStorageLevelConstraint(GenericConstraint):
