+++ zen_garden/elements/technology/constraints/cost_opex_yearly_constraint.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class CostOpexYearlyConstraint(GenericConstraint):
