+++ zen_garden/elements/technology/constraints/technology_capacity_limit_constraint.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class TechnologyCapacityLimitConstraint(GenericConstraint):
