+++ zen_garden/elements/technology/constraints/technology_capacity_lower_limit_constraint.py
@@ -1,4 +1,4 @@
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class TechnologyCapacityLowerLimitConstraint(GenericConstraint):
