+++ zen_garden/elements/technology/constraints/technology_construction_time_constraint.py
@@ -6,7 +6,7 @@ import pandas as pd
 import xarray as xr
 from linopy.expressions import LinearExpression
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class TechnologyConstructionTimeConstraint(GenericConstraint):
