+++ zen_garden/elements/technology/constraints/technology_max_capacity_addition_constraint.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class TechnologyMaxCapacityAdditionConstraint(GenericConstraint):
