+++ zen_garden/elements/technology/constraints/technology_min_capacity_addition_constraint.py
@@ -1,4 +1,4 @@
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class TechnologyMinCapacityAdditionConstraint(GenericConstraint):
