+++ zen_garden/elements/transport_technology/constraints/capacity_factor_transport_constraint.py
@@ -1,6 +1,6 @@
 import xarray as xr
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class CapacityFactorTransportConstraint(GenericConstraint):
