+++ zen_garden/elements/transport_technology/constraints/transport_technology_capex_constraint.py
@@ -1,9 +1,9 @@
 import numpy as np
 import xarray as xr
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
 from zen_garden.elements.transport_technology import TransportTechnology
 from zen_garden.model.components.set_registry import SetRegistry
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class TransportTechnologyCapexConstraint(GenericConstraint):
