+++ zen_garden/elements/transport_technology/constraints/transport_technology_losses_flow_constraint.py
@@ -3,7 +3,7 @@ from typing import cast
 import numpy as np
 import xarray as xr
 
-from zen_garden.constraints.generic_constraint import GenericConstraint
+from zen_garden.topology.generic_constraint import GenericConstraint
 
 
 class TransportTechnologyLossesFlowConstraint(GenericConstraint):
